requirements.txt
diffusers==0.16.1
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
Requests==2.30.0
scipy==1.10.1
tensorflow==2.13.0rc0
tensorflow_macos==2.13.0rc0
torch==2.0.0
transformers==4.28.1
